python-web-aws/requirements.txt
asgiref==3.6.0
boto3==1.26.148
botocore==1.29.148
certifi==2022.12.7
crispy-bootstrap4==2022.1
distlib==0.3.6
Django==4.2
django-appconf==1.0.5
django-crispy-forms==2.0
django-imagekit==4.1.0
django-storages==1.13.2
filelock==3.9.0
fontawesomefree==5.15.4
gunicorn==20.1.0
jmespath==1.0.1
pilkit==2.0
Pillow==9.5.0
pipenv==2023.2.4
platformdirs==3.0.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.19.0
virtualenv-clone==0.5.7
